+++ src/FileTools.java
@@ -12,10 +12,7 @@ import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
-import java.io.IOException;
 import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Locale;
@@ -52,6 +49,9 @@ public class FileTools {
      */
     private src.FileView fileView;
 
+    /**
+     *  Storage for folder path
+     */
     private String folder;
 
     /**
