+++ src/main.java
@@ -13,12 +13,10 @@ class Main {
 	/**
 	 * Main Class that will act as a driver for TeamName project.
 	 *
-	 * @author  TeamName
-	 * @version 1.0
-	 * @since   5/22/2022
-	 */
-
-	/**
+	 * @author  Patrick Tibbals
+	 * @version 1.4
+	 * @since   6/8/2022
+	 *
 	 * Main method calls on guiBuilder method from the MainView class to display the start place
 	 * of our program. Initally the program will load and give the user the option
 	 * to either import/export an item, or sign-on.
