--- src/resources/Exports/Bob_Keener.txt
@@ -1,2 +1,0 @@
-Username,Email,Password,Permissions
-Bob_Keener,Bob@bob.com,**********,true
