--- src/resources/Exports/pa.txt
@@ -1,2 +1,0 @@
-Username,Email,Password,Permissions
-pa,p,p,false
