+++ src/testing/RegistrationTest.java
@@ -8,6 +8,8 @@ import static org.junit.Assert.*;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.net.URISyntaxException;
+
 import org.junit.jupiter.api.Test;
 import src.Registration;
 
@@ -38,9 +40,9 @@ class RegistrationTest {
      * 
      * creates instance of registration to apply test on.
      */
-	public Registration mySubClass() throws FileNotFoundException {
+	public Registration mySubClass() throws FileNotFoundException, URISyntaxException {
 		
-		return new Registration();
+		return new Registration("src/resources/");
 	} 
 	
 	/**
@@ -50,9 +52,12 @@ class RegistrationTest {
 	 * Tests that loginSuccessful returns true with correct username/email/password.
 	 */
 	@Test
-	void testLogInSuccessful() throws FileNotFoundException {
+	void testLogInSuccessful() throws IOException, URISyntaxException {
 		r = mySubClass();
-		assertEquals("Login Should be Successful but is not", r.loginSuccessful("Bob_Keener", "Bob@bob.com", "**********"), true);
+	//	assertEquals("Login Should be Successful but is not", r.loginSuccessful("Bob_Keener", "Bob@bob.com", "**********"), true);
+		r.addToList("joeshmo","shmody@123.com", "768", false);
+		assertEquals("Login Should be Successful but is not", r.loginSuccessful("joeshmo", "shmody@123.com", "768"), true);
+
 	}
 	
 	/**
@@ -62,7 +67,7 @@ class RegistrationTest {
 	 * Tests that loginSuccessful returns false with correct username/email but incorrect password.
 	 */
 	@Test
-	void testLogInSuccessful2() throws FileNotFoundException {
+	void testLogInSuccessful2() throws IOException, URISyntaxException {
 		r = mySubClass();
 		assertEquals("Login Should be Successful but is not", r.loginSuccessful("Bob_Keener", "Bob@bob.com", "zoop"), false);
 	}
@@ -74,7 +79,7 @@ class RegistrationTest {
 	 * Tests that loginSuccessful returns false with correct username/password but incorrect email.
 	 */
 	@Test
-	void testLogInSuccessful3() throws FileNotFoundException {
+	void testLogInSuccessful3() throws FileNotFoundException, URISyntaxException {
 		r = mySubClass();
 		assertEquals("Login Should be Successful but is not", r.loginSuccessful("Bob_Keener", "123@bob.com", "**********"), false);
 	}
@@ -86,7 +91,7 @@ class RegistrationTest {
 	 * Tests that loginSuccessful returns false with correct email/password but incorrect username.
 	 */
 	@Test
-	void testLogInSuccessful4() throws FileNotFoundException {
+	void testLogInSuccessful4() throws FileNotFoundException, URISyntaxException {
 		r = mySubClass();
 		assertEquals("Login Should be Successful but is not", r.loginSuccessful("zerp", "Bob@bob.com", "**********"), false);
 	}
@@ -98,7 +103,7 @@ class RegistrationTest {
 	 * Tests that loginSuccessful returns false with incorrect username/email/password.
 	 */
 	@Test
-	void testLogInSuccessful5() throws FileNotFoundException {
+	void testLogInSuccessful5() throws FileNotFoundException, URISyntaxException {
 		r = mySubClass();
 		assertEquals("Login Should be Successful but is not", r.loginSuccessful("ner", "nerpy@bob.com", "sdlfsldfds"), false);
 	}
@@ -110,7 +115,7 @@ class RegistrationTest {
 	 * Tests that addToList method in registration successfully adds a new user to our arraylist.
 	 * If the user has been added, then the method login successful using the new user information should return true.
 	 */
-	void testAddToList() throws IOException {
+	void testAddToList() throws IOException, URISyntaxException {
 		Registration r = mySubClass();
 		r.addToList("joeshmo", "joeShmo@shmoey", "itsjoe", false);
 		assertEquals("Should add to list of users", r.loginSuccessful("joeshmo", "joeShmo@shmoey", "itsjoe"), true);
